minor optimizations for the new Version RGNDS::GL2D

+++ source/ship.h
@@ -12,7 +12,7 @@
 
 #include "shipengine.h"
 
-#define PointF RGNDS::Point<float>
+#define PointF RGNDS::Point<double>
 
 class Ship : public RGNDS::Transform {
     public:
